requirements.txt
langchain
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit
faiss-cpu==1.7.4
streamlit-extras
altair==4.0
tiktoken
openai
pandas
PyMuPDF
chromadb
pysqlite3-binary
#-e .